requirements.txt
torch
torchvision
tensorflow
numpy
Pillow
matplotlib
scikit-learn
opencv-python